+++ app/controllers/subjects_controller.rb
@@ -3,11 +3,11 @@
 class SubjectsController < ApplicationController
   before_action :authenticate_user!
   before_action :set_viewable_project,                  only: [:index, :show, :timeline, :comments, :settings, :files, :adverse_events, :events, :sheets, :event, :report, :search, :choose_site]
-  before_action :set_editable_project_or_editable_site, only: [:new, :edit, :create, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
-  before_action :redirect_without_project,              only: [:index, :show, :timeline, :comments, :settings, :files, :adverse_events, :events, :sheets, :event, :report, :search, :choose_site, :new, :edit, :create, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
+  before_action :set_editable_project_or_editable_site, only: [:new, :edit, :create, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :set_sheet_as_missing, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
+  before_action :redirect_without_project,              only: [:index, :show, :timeline, :comments, :settings, :files, :adverse_events, :events, :sheets, :event, :report, :search, :choose_site, :new, :edit, :create, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :set_sheet_as_missing, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
   before_action :set_viewable_subject,                  only: [:show, :timeline, :comments, :settings, :files, :adverse_events, :events, :sheets, :event]
-  before_action :set_editable_subject,                  only: [:edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
-  before_action :set_design,                            only: [:new_data_entry]
+  before_action :set_editable_subject,                  only: [:edit, :update, :destroy, :choose_date, :choose_an_event_for_subject, :data_entry, :set_sheet_as_missing, :new_data_entry, :choose_event, :launch_subject_event, :edit_event, :update_event, :destroy_event]
+  before_action :set_design,                            only: [:new_data_entry, :set_sheet_as_missing]
 
   def data_entry
   end
@@ -17,6 +17,12 @@ class SubjectsController < ApplicationController
     # @sheet = @subject.sheets.new(project_id: @project.id, design_id: @design.id, subject_event_id: subject_event_id)
     @sheet = @subject.sheets.where(project_id: @project.id, design_id: @design.id, adverse_event_id: params[:adverse_event_id]).new(sheet_params)
     render 'sheets/new'
+  end
+
+  def set_sheet_as_missing
+    @sheet = @subject.sheets.new(project_id: @project.id, design_id: @design.id, subject_event_id: params[:subject_event_id], missing: true)
+    SheetTransaction.save_sheet!(@sheet, {}, {}, current_user, request.remote_ip, 'sheet_create')
+    redirect_to event_project_subject_path(@project, @subject, event_id: @sheet.subject_event.event, subject_event_id: @sheet.subject_event.id, event_date: @sheet.subject_event.event_date_to_param)
   end
 
   def choose_event