+++ db/migrate/20160219153103_add_missing_to_sheets.rb
@@ -1,0 +1,6 @@
+class AddMissingToSheets < ActiveRecord::Migration
+  def change
+    add_column :sheets, :missing, :boolean, null: false, default: false
+    add_index :sheets, :missing
+  end
+end